+++ labpack/storage/google/drive.py
@@ -3,18 +3,14 @@ __created__ = '2017.07'
 __license__ = 'MIT'
 
 '''
-TODO: handle intermittent connectivity in export/import process
-TODO: handle variable permissions in export/import process
-TODO: add google document export method to load/save process
-TODO: store current folder path ids in import method to reduce redundant requests
-TODO: figure out a way to reduce redundant requests in list method with previous key
-'''
+PLEASE NOTE:    drive package requires the google-api-python-client module.
 
-'''
+(all platforms) pip3 install google-api-python-client
+
 PLEASE NOTE:    google drive is not designed to store files in the POSIX file hierarchy
                 as a result, numerous requests are required to create the parentage
                 and ensure that records exist in the correct nodes. if preserving the
-                parentage does not matter, you can use the method specified here:
+                parentage does not matter, you can use the methods specified here:
                 https://developers.google.com/drive/v3/web/manage-uploads
 
 PLEASE NOTE:    oauth2 tokens granted by google drive expire quickly - usually after
@@ -27,6 +23,19 @@ PLEASE NOTE:    however, to write to a google drive account using a non-approved
                 the oauth2 grantee account must also join this google group
                 https://groups.google.com/forum/#!forum/risky-access-by-unreviewed-apps
 '''
+
+try:
+    from googleapiclient import discovery
+except:
+    import sys
+    print('drive package requires the google-api-python-client module. try: pip3 install google-api-python-client')
+    sys.exit(1)
+
+# TODO: handle intermittent connectivity in export/import process
+# TODO: handle variable permissions in export/import process
+# TODO: add google document export method to load/save process
+# TODO: store current folder path ids in import method to reduce redundant requests
+# TODO: figure out a way to reduce redundant requests in list method with previous key
 
 import os
 from jsonmodel.validators import jsonModel
@@ -1055,84 +1064,3 @@ class driveClient(object):
             skip_insert = ' %s record%s skipped to avoid overwrite.' % (str(skipped), skip_plural)
         exit_msg = '%s record%s exported to %s.%s' % (str(count), plural, new_folder, skip_insert)
         return exit_msg
-            
-if __name__ == '__main__':
-    
-# initialize client
-    import pytest
-    from pprint import pprint
-    from labpack.records.settings import load_settings
-    google_tokens = load_settings('../../../../cred/tokens/google-drive.yaml')
-    access_token = google_tokens['google_drive_access_token']
-    drive_client = driveClient(access_token, 'Unit Test')
-
-# prevent accidental use
-    assert drive_client.drive_space == 'appDataFolder'
-    count = 0
-    for id, name, mimetype in drive_client._list_directory():
-        count += 1
-        break
-    assert not count
-
-# construct test records
-    import json
-    from hashlib import md5
-    from labpack.compilers import drep
-    secret_key = 'upside'
-    test_record = {
-        'dt': 1474509314.419702,
-        'deviceID': '2Pp8d9lpsappm8QPv_Ps6cL0'
-    }
-    test_data = open('../../../data/test_voice.ogg', 'rb').read()
-    data_key = 'lab/voice/unittest.ogg'
-    record_data = json.dumps(test_record).encode('utf-8')
-    record_key = 'lab/device/unittest.json'
-    drep_data = drep.dump(test_record, secret_key)
-    drep_key = 'lab/device/unittest.drep'
-
-# test save method
-    old_hash = md5(test_data).digest()
-    drive_client.save(data_key, test_data, secret_key=secret_key)
-    drive_client.save(record_key, record_data)
-    drive_client.save(drep_key, drep_data)
-    assert drive_client.exists(drep_key)
-    assert not drive_client.exists('notakey')
-
-# test list with different filters
-    data_filter = { 2:{'must_contain': ['ogg$']}}
-    filter_function = drive_client.conditional_filter(data_filter)
-    data_search = drive_client.list(filter_function=filter_function, max_results=3)
-    prefix_search = drive_client.list(prefix='lab/dev', delimiter='drep', max_results=3)
-    print(prefix_search)
-
-# test import and export method
-    try:
-        from labpack.storage.appdata import appdataClient
-        export_client = appdataClient(collection_name='Test Export')
-        drive_client.export(export_client)
-        export_status = drive_client.export(export_client, overwrite=False)
-        print(export_status)
-        export_list = export_client.list(filter_function=filter_function, max_results=3)
-        print(export_list)
-        import_status = export_client.export(drive_client)
-        print(import_status)
-        export_client.remove()
-    except Exception as err:
-        print(err)
-
-# test load method
-    new_data = drive_client.load(data_search[0], secret_key=secret_key)
-    new_hash = md5(new_data).digest()
-    assert old_hash == new_hash
-    load_data = drive_client.load(prefix_search[0])
-    record_details = json.loads(load_data.decode())
-    assert record_details == test_record
-    with pytest.raises(Exception):
-        drive_client.load('notakey')
-
-# test delete
-    delete_status = drive_client.delete(data_key)
-
-# test removal
-    remove_status = drive_client.remove()
-    print(remove_status)